requirements.txt
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyaml==24.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
seaborn==0.13.2
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
zipp==3.19.2
